requirements.txt
streamlit
python-dotenv
google-genai
fastapi
uvicorn
pydantic
requests
pandas
numpy
sqlalchemy
email-validator
neo4j